Refactored And Added Date Range Picker
Further refactored the Admin Dashboard file structure to match the regular Dashboard file structure, and resolved build errors.

Also updated the Date Picker to be a common component for reuse in the Admin Dashboard as well as the existing Dashboard
