+++ src/app/admin/page.tsx
@@ -1,5 +1,5 @@
-import { AdminDashboardPage } from '@/components/pages/AdminDashboardPage'
+import AdminDashboardPage from '@/components/pages/AdminDashboardPage';
 
-export default function AdminRoute() {
-  return <AdminDashboardPage />
+export default function AdminPage() {
+  return <AdminDashboardPage />;
 }
