+++ src/components/layouts/AdminDashboardPageLayout/AdminDashboardPageLayoutClient.tsx
@@ -4,25 +4,46 @@ import { User } from '@supabase/supabase-js';
 import { Typography } from '@/components/ui/typography';
 import { TabNavigation, TabItem } from '@/components/common/TabNavigation';
 import { usePathname, useSearchParams } from 'next/navigation';
-import { useMemo } from 'react';
+import { useMemo, useState, createContext } from 'react';
+import { DateRange } from 'react-day-picker';
+import { DateRangePicker } from '@/components/common/DateRangePicker/DateRangePicker';
+import React from 'react';
+
+export type DateRangeContextType = {
+  dateRange: DateRange | undefined;
+  setDateRange: (range: DateRange | undefined) => void;
+};
+
+export const DateRangeContext = createContext<DateRangeContextType | undefined>(undefined);
 
 export type AdminDashboardPageLayoutClientProps = {
   user: User;
   children: React.ReactNode;
+  dashboardData: {
+    totalBookings: number;
+    bookingsPerDay: Record<string, number>;
+    totalClients: number;
+    newClients: number;
+    totalProfessionals: number;
+    newProfessionals: number;
+    totalChats: number;
+    totalRefunds: number;
+  };
 };
 
-export function AdminDashboardPageLayoutClient({ user, children }: AdminDashboardPageLayoutClientProps) {
+export function AdminDashboardPageLayoutClient({ user, children, dashboardData }: AdminDashboardPageLayoutClientProps) {
   const pathname = usePathname();
   const searchParams = useSearchParams();
+  const [dateRange, setDateRange] = useState<DateRange | undefined>();
 
   // Determine active tab from pathname
   const getActiveTabFromPath = (path: string): string => {
-    if (path === '/admin/dashboard' || path === '/admin/dashboard/') return 'overview';
-    if (path.includes('/admin/dashboard/bookings')) return 'bookings';
-    if (path.includes('/admin/dashboard/clients')) return 'clients';
-    if (path.includes('/admin/dashboard/professionals')) return 'professionals';
-    if (path.includes('/admin/dashboard/refunds')) return 'refunds';
-    if (path.includes('/admin/dashboard/messages')) return 'messages';
+    if (path === '/admin' || path === '/admin/') return 'overview';
+    if (path.includes('/admin/bookings')) return 'bookings';
+    if (path.includes('/admin/clients')) return 'clients';
+    if (path.includes('/admin/professionals')) return 'professionals';
+    if (path.includes('/admin/refunds')) return 'refunds';
+    if (path.includes('/admin/messages')) return 'messages';
     return 'overview';
   };
 
@@ -35,45 +56,66 @@ export function AdminDashboardPageLayoutClient({ user, children }: AdminDashboar
   };
 
   const tabs: TabItem[] = useMemo(() => [
-    { key: 'overview', label: 'Overview', href: createTabUrl('/admin/dashboard'), isActive: activeTab === 'overview' },
-    { key: 'bookings', label: 'Bookings', href: createTabUrl('/admin/dashboard/bookings'), isActive: activeTab === 'bookings' },
-    { key: 'clients', label: 'Clients', href: createTabUrl('/admin/dashboard/clients'), isActive: activeTab === 'clients' },
-    { key: 'professionals', label: 'Professionals', href: createTabUrl('/admin/dashboard/professionals'), isActive: activeTab === 'professionals' },
-    { key: 'refunds', label: 'Refunds', href: createTabUrl('/admin/dashboard/refunds'), isActive: activeTab === 'refunds' },
-    { key: 'messages', label: 'Messages', href: createTabUrl('/admin/dashboard/messages'), isActive: activeTab === 'messages' },
+    { key: 'overview', label: 'Overview', href: createTabUrl('/admin'), isActive: activeTab === 'overview' },
+    { key: 'bookings', label: 'Bookings', href: createTabUrl('/admin/bookings'), isActive: activeTab === 'bookings' },
+    { key: 'clients', label: 'Clients', href: createTabUrl('/admin/clients'), isActive: activeTab === 'clients' },
+    { key: 'professionals', label: 'Professionals', href: createTabUrl('/admin/professionals'), isActive: activeTab === 'professionals' },
+    { key: 'refunds', label: 'Refunds', href: createTabUrl('/admin/refunds'), isActive: activeTab === 'refunds' },
+    { key: 'messages', label: 'Messages', href: createTabUrl('/admin/messages'), isActive: activeTab === 'messages' },
   ], [activeTab, searchParams]);
 
   const displayName = [user.user_metadata?.first_name, user.user_metadata?.last_name]
     .filter(Boolean)
     .join(' ') || user.email;
 
+  // Propagate dashboardData to children if they are valid React elements
+  const childrenWithDashboardData = React.Children.map(children, child => {
+    if (
+      React.isValidElement(child) &&
+      typeof child.type === 'function' &&
+      !('dashboardData' in (child.props as object))
+    ) {
+      return React.cloneElement(
+        child as React.ReactElement<{ dashboardData: typeof dashboardData }>,
+        { dashboardData }
+      );
+    }
+    return child;
+  });
+
   return (
-    <div className="w-full mx-auto space-y-4 lg:space-y-6">
-      {/* Page header */}
-      <div className="flex flex-col sm:flex-row justify-between items-start gap-3 lg:gap-4">
-        <div className="flex-1 min-w-0">
-          <Typography
-            variant="h2"
-            className="leading-5 border-none font-bold text-foreground text-xl lg:text-2xl"
-          >
-            Admin Dashboard
-          </Typography>
-          <Typography className="text-muted-foreground text-sm lg:text-base truncate">
-            Welcome, {displayName}
-          </Typography>
+    <DateRangeContext.Provider value={{ dateRange, setDateRange }}>
+      <div className="w-full mx-auto space-y-4 lg:space-y-6">
+        {/* Page header */}
+        <div className="flex flex-col sm:flex-row justify-between items-start gap-3 lg:gap-4">
+          <div className="flex-1 min-w-0">
+            <Typography
+              variant="h2"
+              className="leading-5 border-none font-bold text-foreground text-xl lg:text-2xl"
+            >
+              Admin Dashboard
+            </Typography>
+            <Typography className="text-muted-foreground text-sm lg:text-base truncate">
+              Welcome, {displayName}
+            </Typography>
+          </div>
         </div>
-        {/* Placeholder for future date picker or admin actions */}
+        {/* Tab navigation and date picker row */}
+        <div className="flex flex-col md:flex-row items-stretch md:items-center justify-between gap-2 mb-4 lg:mb-6">
+          <div className="overflow-x-auto flex-1 min-w-0">
+            <TabNavigation
+              tabs={tabs}
+              variant="link"
+              className="min-w-max"
+            />
+          </div>
+          <div className="w-full md:w-auto md:ml-4 flex-shrink-0">
+            <DateRangePicker dateRange={dateRange} onDateRangeChange={setDateRange} />
+          </div>
+        </div>
+        {/* Main content */}
+        <div className="mt-4">{childrenWithDashboardData}</div>
       </div>
-      {/* Tab navigation */}
-      <div className="overflow-x-auto">
-        <TabNavigation
-          tabs={tabs}
-          variant="link"
-          className="mb-4 lg:mb-6 min-w-max"
-        />
-      </div>
-      {/* Main content */}
-      <div className="mt-4">{children}</div>
-    </div>
+    </DateRangeContext.Provider>
   );
 }
