+++ src/components/pages/AdminDashboardPage/index.ts
@@ -1,1 +1,1 @@
-export { AdminDashboardPage } from './AdminDashboardPage' 
+export { default } from './AdminDashboardPage';
